@@ -49,84 +51,85 @@
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slider": "^1.1.2",
     "@radix-ui/react-tabs": "^1.0.4",
-    "@reduxjs/toolkit": "^1.9.7",
-    "@storybook/test": "8.0.8",
+    "@reduxjs/toolkit": "^2.2.5",
     "@svgr/webpack": "^8.1.0",
     "@types/cors": "^2.8.17",
-    "@types/react-avatar-editor": "^13.0.2",
-    "@types/react-dom": "^18.2.25",
-    "@types/react-slick": "^0.23.13",
+    "async-mutex": "^0.5.0",
     "axios": "^1.6.8",
-    "clsx": "^2.1.0",
+    "clsx": "^2.1.1",
     "cors": "^2.8.5",
     "date-fns": "^3.6.0",
-    "javascript-time-ago": "^2.5.9",
-    "next": "14.2.0",
+    "javascript-time-ago": "^2.5.10",
+    "next": "14.2.3",
     "next-redux-wrapper": "^8.1.0",
     "nprogress": "^0.2.0",
-    "react": "18.2.0",
+    "react": "18.3.1",
     "react-avatar-editor": "^13.0.2",
-    "react-datepicker": "^4.25.0",
-    "react-dom": "18.2.0",
-    "react-easy-crop": "^5.0.6",
-    "react-hook-form": "^7.51.3",
-    "react-redux": "^9.1.0",
-    "react-slick": "^0.29.0",
+    "react-datepicker": "^6.9.0",
+    "react-dom": "18.3.1",
+    "react-easy-crop": "^5.0.7",
+    "react-hook-form": "^7.51.4",
+    "react-redux": "^9.1.2",
+    "react-slick": "^0.30.2",
     "react-toastify": "^10.0.5",
-    "slick-carousel": "^1.8.1",
     "typescript": "^5.4.5",
     "use-debounce": "^10.0.0",
-    "zod": "^3.22.4"
+    "zod": "^3.22.8"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^1.3.2",
+    "@chromatic-com/storybook": "^1.4.0",
     "@conarti/eslint-plugin-feature-sliced": "^1.0.5",
     "@hookform/devtools": "^4.3.1",
     "@it-incubator/eslint-config": "^1.0.2",
     "@it-incubator/prettier-config": "^0.1.2",
-    "@it-incubator/stylelint-config": "^1.0.3",
-    "@next/bundle-analyzer": "^14.2.0",
-    "@next/eslint-plugin-next": "^14.2.0",
-    "@playwright/test": "^1.43.0",
-    "@storybook/addon-designs": "^8.0.0",
-    "@storybook/addon-essentials": "^8.0.8",
-    "@storybook/addon-interactions": "^8.0.8",
-    "@storybook/addon-links": "^8.0.8",
-    "@storybook/addon-onboarding": "^8.0.8",
-    "@storybook/blocks": "^8.0.8",
-    "@storybook/nextjs": "^8.0.8",
-    "@storybook/react": "^8.0.8",
-    "@swc/core": "^1.4.13",
+    "@it-incubator/stylelint-config": "^2.0.0",
+    "@next/bundle-analyzer": "^14.2.3",
+    "@next/eslint-plugin-next": "^14.2.3",
+    "@playwright/test": "^1.44.0",
+    "@storybook/addon-actions": "^8.1.1",
+    "@storybook/addon-designs": "^8.0.1",
+    "@storybook/addon-essentials": "^8.1.1",
+    "@storybook/addon-interactions": "^8.1.1",
+    "@storybook/addon-links": "^8.1.1",
+    "@storybook/addon-onboarding": "^8.1.1",
+    "@storybook/blocks": "^8.1.1",
+    "@storybook/nextjs": "^8.1.1",
+    "@storybook/react": "^8.1.1",
+    "@storybook/test": "^8.1.1",
+    "@swc/core": "^1.5.7",
     "@swc/jest": "^0.2.36",
-    "@testing-library/dom": "^9.3.4",
-    "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^14.3.1",
+    "@testing-library/dom": "^10.1.0",
+    "@testing-library/jest-dom": "^6.4.5",
+    "@testing-library/react": "^15.0.7",
     "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.12",
-    "@types/node": "^20.12.7",
+    "@types/node": "^20.12.12",
     "@types/nprogress": "^0.2.3",
-    "@types/react": "18.2.77",
-    "@types/react-datepicker": "^4.19.6",
+    "@types/react": "18.3.2",
+    "@types/react-avatar-editor": "^13.0.2",
+    "@types/react-datepicker": "^6.2.0",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-slick": "^0.23.13",
     "@welldone-software/why-did-you-render": "^8.0.1",
     "cross-env": "^7.0.3",
     "cross-fetch": "^4.0.0",
     "eslint": "^8.57.0",
-    "eslint-config-next": "^14.2.0",
+    "eslint-config-next": "^14.2.3",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^28.2.0",
+    "eslint-plugin-jest": "^28.5.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-storybook": "^0.8.0",
-    "eslint-plugin-testing-library": "^6.2.0",
+    "eslint-plugin-testing-library": "^6.2.2",
     "husky": "^9.0.11",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "matchmedia-polyfill": "^0.3.2",
     "next-router-mock": "^0.9.13",
-    "react-test-renderer": "^18.2.0",
-    "sass": "^1.75.0",
-    "storybook": "^8.0.8",
-    "stylelint": "^16.3.1",
+    "react-test-renderer": "^18.3.1",
+    "sass": "^1.77.1",
+    "storybook": "^8.1.1",
+    "stylelint": "^16.5.0",
     "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2"
   }
