@@ -1,5 +1,6 @@
 import { SingUpForm } from '@/features/auth/signUp/ui/singUpForm'
 import { Meta, StoryObj } from '@storybook/react'
+import { fn } from '@storybook/test'
 
 const meta = {
   argTypes: {
@@ -13,8 +14,10 @@ const meta = {
 } satisfies Meta<typeof SingUpForm>
 
 export default meta
-type Story = StoryObj<typeof SingUpForm>
+type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-  args: {},
+  args: {
+    onSubmitHandler: fn(),
+  },
 }
