@@ -11,15 +11,15 @@ import { NextPage } from 'next'
 import '@/styles/_globals.scss'
 import '@/styles/nprogress.scss'
 import 'react-toastify/dist/ReactToastify.css'
-import '../scripts/wdyr'
+// import '../scripts/wdyr' необходимо для просмотра что именно рендериться повторно
 
-type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
+type NextPageWithLayout<P = {}, IP = P> = {
   getLayout?: (page: ReactElement) => ReactNode
-}
+} & NextPage<P, IP>
 
-type AppPropsWithLayout = AppProps & {
+type AppPropsWithLayout = {
   Component: NextPageWithLayout
-}
+} & AppProps
 
 export default function App({ Component, ...rest }: AppPropsWithLayout) {
   const { props, store } = wrapper.useWrappedStore(rest)
