@@ -1,4 +1,4 @@
-import { memo, useCallback, useMemo } from 'react'
+import { memo, useMemo } from 'react'
 
 import FlagRussiaIcon from '@/public/icon/flagRussiaIcon.svg'
 import FlagUKIcon from '@/public/icon/flagUKIcon.svg'
@@ -6,10 +6,10 @@ import OutlineBellIcon from '@/public/icon/outlineBellIcon.svg'
 import { RouteNames } from '@/shared/const'
 import { useTranslate } from '@/shared/hooks'
 import { Button } from '@/ui/button'
-import { Option, SelectBox } from '@/ui/selectBox'
 import { Typography } from '@/ui/typography'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
+import { Option, Select } from 'src/shared/ui/select'
 
 import s from '@/shared/header/ui/header.module.scss'
 
@@ -27,16 +27,11 @@ export const Header = memo(function Header({ variant }: Props) {
       { id: 'rus01', image: <FlagRussiaIcon />, value: 'Russian' },
     ]
   }, [])
-  const changeLangHandler = useCallback(
-    (value: number | string) => {
-      if (typeof value == 'string') {
-        const locale = value.slice(0, 2).toLowerCase()
+  const changeLangHandler = (value: string) => {
+    const locale = value.slice(0, 2).toLowerCase()
 
-        push({ pathname, query }, asPath, { locale })
-      }
-    },
-    [asPath, pathname, push, query]
-  )
+    push({ pathname, query }, asPath, { locale })
+  }
   const defaultValue = useMemo(() => {
     return locale === 'en' ? languages[0].value : languages[1].value
   }, [languages, locale])
@@ -45,17 +40,18 @@ export const Header = memo(function Header({ variant }: Props) {
     <header className={s.header}>
       <div className={s.content}>
         <Typography className={s.text} variant={'large'}>
-          <Link className={s.incta} href={'/'} tabIndex={1}>
+          <Link className={s.incta} href={'/'} tabIndex={0}>
             Inсtagram
           </Link>
         </Typography>
         <div className={s.options_container}>
           <OutlineBellIcon />
           <div className={s.select}>
-            <SelectBox
-              defaultValue={defaultValue}
+            <Select
               onValueChange={changeLangHandler}
               options={languages}
+              placeholder={t.locale.language}
+              // value={defaultValue}
             />
           </div>
           {variant === 'public' && (
