@@ -1,3 +1,9 @@
 import { StateSchema } from '@/store'
+import { createSelector } from '@reduxjs/toolkit'
 
-export const getUserId = (state: StateSchema) => state.authMe.authMeData?.userId ?? state.signIn.id
+export const getUserIdMemo = (state: StateSchema) => state.authMe.authMeData?.userId
+export const getSignInIdMemo = (state: StateSchema) => state.signIn.id
+
+export const getUserId = createSelector([getUserIdMemo, getSignInIdMemo], (userId, id) => {
+  return userId ?? id
+})
