@@ -1,3 +1,9 @@
 import { StateSchema } from '@/store'
+import { createSelector } from '@reduxjs/toolkit'
 
-export const profileFoundSelector = (state: StateSchema) => state.menu.profileFound
+export const profileFoundSelector = createSelector(
+  [(state: StateSchema) => state.menu.profileFound],
+  profileFound => {
+    return profileFound
+  }
+)
