@@ -1,50 +1,36 @@
-import type { StoryObj } from '@storybook/react'
-
-import { useState } from 'react'
+import type { Meta, StoryObj } from '@storybook/react'
 
 import { Checkbox } from '@/ui/checkbox'
+import { action } from '@storybook/addon-actions'
 
 const meta = {
-  argTypes: {
-    checked: {
-      control: { type: 'boolean' },
-      options: [true, false],
-    },
+  args: {
+    label: 'Check-box',
+    onCheckedChange: action('onChecked'),
   },
   component: Checkbox,
+  tags: ['autodocs'],
   title: 'Components/Checkbox',
-}
+} satisfies Meta<typeof Checkbox>
 
 export default meta
-type Story = StoryObj<typeof Checkbox>
+type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-  render: args => {
-    const [checked, setChecked] = useState(false)
-
-    return (
-      <>
-        <Checkbox
-          {...args}
-          checked={checked}
-          disabled={false}
-          label={'Check-box'}
-          onChange={() => setChecked(!checked)}
-        />
-      </>
-    )
-  },
+  args: {},
 }
 export const Checked: Story = {
-  args: { checked: true, disabled: false, label: 'Check-box' },
+  args: { checked: true },
 }
 export const NotChecked: Story = {
-  args: { checked: false, disabled: false, label: 'Check-box' },
+  args: { checked: false },
 }
 export const DisabledAndChecked: Story = {
-  args: { checked: true, disabled: true, label: 'Check-box' },
+  args: {
+    checked: true,
+    disabled: true,
+  },
 }
-
 export const DisabledAndNotChecked: Story = {
-  args: { checked: false, disabled: true, label: 'Check-box' },
+  args: { checked: false, disabled: true },
 }
