@@ -1,4 +1,4 @@
-import { ComponentProps, forwardRef } from 'react'
+import { ComponentPropsWithoutRef, SyntheticEvent, forwardRef } from 'react'
 import * as RDP from 'react-datepicker'
 import { ReactDatePickerCustomHeaderProps } from 'react-datepicker'
 import { FieldValues } from 'react-hook-form'
@@ -10,9 +10,10 @@ import { RouteNames } from '@/shared/const'
 import { useTranslate } from '@/shared/hooks'
 import { Label } from '@/ui/label'
 import { Typography } from '@/ui/typography'
+import { offset } from '@floating-ui/dom'
 import { clsx } from 'clsx'
 // eslint-disable-next-line import/no-duplicates
-import { format } from 'date-fns'
+import { Locale, format } from 'date-fns'
 // eslint-disable-next-line import/no-duplicates
 import { enGB, ru } from 'date-fns/locale'
 import { useRouter } from 'next/router'
@@ -25,14 +26,14 @@ import textFieldStyles from '@/ui/textField/textField.module.scss'
 export type DatePickerProps = {
   disabled?: boolean
   endDate?: Date | null
-  errorMessage?: string
+  error?: string
   label?: string
   onChange?: (value: Date | Date[] | null) => void
   placeholder?: string
   setEndDate?: (date: Date | null) => void
   setStartDate: (date: Date | null) => void
   startDate: Date | null
-} & ComponentProps<'div'>
+} & ComponentPropsWithoutRef<'div'>
 
 const RDPC = (((RDP.default as any).default as any) ||
   (RDP.default as any) ||
@@ -44,7 +45,7 @@ export const DatePicker = forwardRef<FieldValues, DatePickerProps>(
       className,
       disabled,
       endDate,
-      errorMessage,
+      error,
       label,
       placeholder,
       setEndDate,
@@ -56,7 +57,7 @@ export const DatePicker = forwardRef<FieldValues, DatePickerProps>(
   ) => {
     const isRange = endDate !== undefined
 
-    const showError = !!errorMessage && errorMessage.length > 0
+    const showError = !!error && error.length > 0
     const { t } = useTranslate()
     const router = useRouter()
 
@@ -70,7 +71,10 @@ export const DatePicker = forwardRef<FieldValues, DatePickerProps>(
       root: clsx(s.root, className),
     }
 
-    const DatePickerHandler = (dates: [Date | null, Date | null] | Date) => {
+    const DatePickerHandler = (
+      dates: Date | null,
+      event: SyntheticEvent<any, Event> | undefined
+    ) => {
       if (Array.isArray(dates)) {
         const [start, end] = dates
 
@@ -81,8 +85,8 @@ export const DatePicker = forwardRef<FieldValues, DatePickerProps>(
       }
     }
     const isError =
-      errorMessage?.includes('A user under 13 cannot create a profile.') ||
-      errorMessage?.includes('Пользователь младше 13 лет не может создать профиль.')
+      error?.includes('A user under 13 cannot create a profile.') ||
+      error?.includes('Пользователь младше 13 лет не может создать профиль.')
 
     return (
       <div className={classNames.root} {...rest}>
@@ -90,7 +94,7 @@ export const DatePicker = forwardRef<FieldValues, DatePickerProps>(
           calendarClassName={classNames.calendar}
           calendarStartDay={1}
           className={classNames.input}
-          customInput={<CustomInput disabled={disabled} error={errorMessage} label={label} />}
+          customInput={<CustomInput disabled={disabled} error={error} label={label} />}
           dateFormat={'dd/MM/yyyy'}
           dayClassName={classNames.day}
           disabled={disabled}
@@ -101,12 +105,10 @@ export const DatePicker = forwardRef<FieldValues, DatePickerProps>(
           placeholderText={placeholder}
           popperClassName={classNames.popper}
           popperModifiers={[
-            {
-              name: 'offset',
-              options: {
-                offset: [0, -11],
-              },
-            },
+            offset({
+              crossAxis: 74,
+              mainAxis: -10,
+            }),
           ]}
           renderCustomHeader={CustomHeaderWrapper(router.locale === 'en' ? enGB : ru)}
           selected={startDate}
@@ -187,7 +189,7 @@ const CustomHeaderWrapper = (locale: Locale) => {
       header: s.header,
     }
 
-    const headerText = capitalizeFirstLetter(format(date, 'LLLL Y', { locale: locale }))
+    const headerText = capitalizeFirstLetter(format(date, 'LLLL y', { locale: locale }))
 
     return (
       <div className={classNames.header}>
