@@ -1,4 +1,5 @@
 import * as React from 'react'
+import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react'
 
 import { Typography } from '@/ui/typography'
 import * as RadioGroup from '@radix-ui/react-radio-group'
@@ -6,18 +7,16 @@ import { clsx } from 'clsx'
 
 import s from '@/ui/radioButton/radioButton.module.scss'
 
-const RadioGroupRoot = React.forwardRef<
-  React.ElementRef<typeof RadioGroup.Root>,
-  React.ComponentPropsWithoutRef<typeof RadioGroup.Root>
+const RadioGroupRoot = forwardRef<
+  ElementRef<typeof RadioGroup.Root>,
+  ComponentPropsWithoutRef<typeof RadioGroup.Root>
 >(({ className, ...props }, ref) => {
   return <RadioGroup.Root className={clsx(s.root, className)} {...props} ref={ref} />
 })
 
-RadioGroupRoot.displayName = RadioGroup.Root.displayName
-
-const RadioGroupItem = React.forwardRef<
-  React.ElementRef<typeof RadioGroup.Item>,
-  React.ComponentPropsWithoutRef<typeof RadioGroup.Item>
+const RadioGroupItem = forwardRef<
+  ElementRef<typeof RadioGroup.Item>,
+  ComponentPropsWithoutRef<typeof RadioGroup.Item>
 >(({ children, className, ...props }, ref) => {
   return (
     <RadioGroup.Item className={clsx(s.option, className)} ref={ref} {...props}>
@@ -26,35 +25,29 @@ const RadioGroupItem = React.forwardRef<
   )
 })
 
-RadioGroupItem.displayName = RadioGroup.Item.displayName
-
 type Option = {
   label: string
   value: string
 }
-export type RadioButtonProps = Omit<
-  React.ComponentPropsWithoutRef<typeof RadioGroup.Root>,
-  'children'
-> & {
+export type RadioButtonProps = {
   error?: string
   options: Option[]
-}
-export const RadioButton = React.forwardRef<
-  React.ElementRef<typeof RadioGroup.Root>,
-  RadioButtonProps
->((props, ref) => {
-  const { error, options, ...restProps } = props
+} & Omit<ComponentPropsWithoutRef<typeof RadioGroup.Root>, 'children'>
+export const RadioButton = forwardRef<ElementRef<typeof RadioGroup.Root>, RadioButtonProps>(
+  (props, ref) => {
+    const { error, options, ...restProps } = props
 
-  return (
-    <RadioGroupRoot {...restProps} ref={ref}>
-      {options.map(option => (
-        <div className={s.label} key={option.value}>
-          <RadioGroupItem id={option.value} value={option.value} />
-          <Typography as={'label'} htmlFor={option.value} variant={'regular14'}>
-            {option.label}
-          </Typography>
-        </div>
-      ))}
-    </RadioGroupRoot>
-  )
-})
+    return (
+      <RadioGroupRoot {...restProps} ref={ref}>
+        {options.map(option => (
+          <div className={s.label} key={option.value}>
+            <RadioGroupItem id={option.value} value={option.value} />
+            <Typography as={'label'} htmlFor={option.value} variant={'regular14'}>
+              {option.label}
+            </Typography>
+          </div>
+        ))}
+      </RadioGroupRoot>
+    )
+  }
+)
