@@ -1,4 +1,4 @@
-import type { StoryObj } from '@storybook/react'
+import type { Meta, StoryObj } from '@storybook/react'
 
 import { RadioButton } from '@/ui/radioButton'
 
@@ -6,10 +6,10 @@ const meta = {
   component: RadioButton,
   tags: ['autodocs'],
   title: 'Components/RadioButton',
-}
+} satisfies Meta<typeof RadioButton>
 
 export default meta
-type Story = StoryObj<typeof RadioButton>
+type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
   args: {
@@ -19,9 +19,19 @@ export const Default: Story = {
     ],
   },
 }
+export const DefaultOneValue: Story = {
+  args: {
+    defaultValue: 'option-one',
+    options: [
+      { label: 'Option One', value: 'option-one' },
+      { label: 'Option Two', value: 'option-two' },
+    ],
+  },
+}
 
 export const Disabled: Story = {
   args: {
+    defaultValue: 'option-two',
     disabled: true,
     options: [
       { label: 'Option One', value: 'option-one' },
