@@ -1,2 +1,2 @@
 export { TextAreaField } from '@/ui/textAreaField/TextAreaField'
-export type { TTextAreaProps } from '@/ui/textAreaField/TextAreaField'
+export type { TextAreaFieldProps } from '@/ui/textAreaField/TextAreaField'
