@@ -1,4 +1,10 @@
-import React, { ChangeEvent, ComponentProps, useState } from 'react'
+import React, {
+  ChangeEvent,
+  ComponentPropsWithoutRef,
+  ElementRef,
+  forwardRef,
+  useState,
+} from 'react'
 
 import EyeIcon from '@/public/icon/eyeIcon.svg'
 import EyeOffIcon from '@/public/icon/eyeOffIcon.svg'
@@ -8,81 +14,69 @@ import { Typography } from '@/ui/typography'
 import s from '@/ui/textField/inputMain/inputMain.module.scss'
 
 type Props = {
-  errorMessage?: string
+  error?: string
+  id?: string
   isRequired?: boolean
   label?: string
   onChangeText?: (value: string) => void
-  value?: string
-} & ComponentProps<'input'>
+} & ComponentPropsWithoutRef<'input'>
 
-export const InputMain = ({
-  disabled,
-  errorMessage,
-  isRequired,
-  label,
-  name,
-  onChangeText,
-  placeholder,
-  type = 'text',
-  value,
-  ...rest
-}: Props) => {
-  const [showPassword, setShowPassword] = useState<boolean>(false)
+export const InputMain = forwardRef<ElementRef<'input'>, Props>(
+  ({ disabled, error, id, isRequired, label, name, onChangeText, type = 'text', ...rest }, ref) => {
+    const [showPassword, setShowPassword] = useState(false)
 
-  const onClickShowValue = () => {
-    if (!disabled) {
-      setShowPassword(!showPassword)
+    const onClickShowValue = () => {
+      if (!disabled) {
+        setShowPassword(!showPassword)
+      }
     }
-  }
 
-  const onchangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
-    onChangeText && onChangeText(e.currentTarget.value)
-  }
+    const onchangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
+      onChangeText?.(e.currentTarget.value)
+    }
 
-  const showErrorMess = errorMessage && errorMessage.length > 0
+    const showErrorMess = error && error.length > 0
 
-  return (
-    <div className={s.textFieldWrap}>
-      <label
-        className={`${s.label} ${disabled && s.disabledLabel}`}
-        data-testid={name}
-        htmlFor={name}
-      >
-        {label}
-        {isRequired && <span className={s.star}>*</span>}
-      </label>
-      <input
-        disabled={disabled}
-        id={name}
-        name={name}
-        onChange={onchangeHandler}
-        placeholder={placeholder && placeholder}
-        role={name}
-        type={showPassword ? 'text' : type}
-        value={value}
-        {...rest}
-        className={`${s.textField} ${errorMessage && s.errorInput} ${disabled && s.disabledInput}`}
-      />
-
-      {type === 'password' && !showPassword && (
-        <EyeIcon
-          className={`${s.iconEye} ${disabled && s.disabledIconEye}`}
-          onClick={onClickShowValue}
+    return (
+      <div className={s.textFieldWrap}>
+        <label
+          className={`${s.label} ${disabled && s.disabledLabel}`}
+          data-testid={name}
+          htmlFor={id}
+        >
+          {label}
+          {isRequired && <span className={s.star}>*</span>}
+        </label>
+        <input
+          id={id}
+          onChange={onchangeHandler}
+          ref={ref}
+          role={name}
+          type={showPassword ? 'text' : type}
+          {...rest}
+          className={`${s.textField} ${error && s.errorInput} ${disabled && s.disabledInput}`}
         />
-      )}
-      {type === 'password' && showPassword && (
-        <EyeOffIcon
-          className={`${s.iconEye} ${disabled && s.disabledIconEye}`}
-          onClick={onClickShowValue}
-        />
-      )}
 
-      {type === 'search' && <SearchIcon className={s.searchIcon} />}
-      {showErrorMess && (
-        <Typography className={s.errorWrap} variant={'regular14'}>
-          {errorMessage}
-        </Typography>
-      )}
-    </div>
-  )
-}
+        {type === 'password' && !showPassword && (
+          <EyeIcon
+            className={`${s.iconEye} ${disabled && s.disabledIconEye}`}
+            onClick={onClickShowValue}
+          />
+        )}
+        {type === 'password' && showPassword && (
+          <EyeOffIcon
+            className={`${s.iconEye} ${disabled && s.disabledIconEye}`}
+            onClick={onClickShowValue}
+          />
+        )}
+
+        {type === 'search' && <SearchIcon className={s.searchIcon} />}
+        {showErrorMess && (
+          <Typography className={s.errorWrap} variant={'regular14'}>
+            {error}
+          </Typography>
+        )}
+      </div>
+    )
+  }
+)
